Make process termination and rlimit setting be cross-plat
